[dev] move 1822CA game state specs to their own file
* use `fixture_at_action()`
* remove `raw_action()` helper function

With the caching from #11970, these tests have noticeable performance
improvements.

Before:

`docker compose exec rack bundle exec rspec spec/game_state_spec.rb -e 1822CA`
takes 2.42 seconds

After:

`docker compose exec rack bundle exec rspec spec/lib/engine/game/g_1822_ca/game_spec.rb`
takes 1.54 seconds
